Cleaned the code 1(moved everything that's related to menus from Core.js into MenuManager.js)
